feat: use typst-assets crate to get fonts
