+++ crates/bevy_typst/Cargo.toml
@@ -10,6 +10,7 @@ bevy_vello = "0.3"
 # typst dependencies
 typst = "0.11"
 typst-svg = "0.11"
+typst-assets = { version = "0.11", optional = true }
 comemo = "0.4" # in sync with typst
 fontdb = { version = "0.16", features = ["memmap", "fontconfig"] }
 chrono = { version = "0.4.24", default-features = false, features = ["clock", "std"] }
@@ -26,4 +27,4 @@ workspace = true
 
 [features]
 default = ["embed-fonts"]
-embed-fonts = []
+embed-fonts = ["dep:typst-assets"]
